# apps/backup-operator/api/openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Supabase Backup Operator API
  version: 1.0.0
security:
  - serviceAssertion: []
paths:
...
  /v1/clusters/{clusterId}/backups:
    get:
      operationId: listClusterBackups
      parameters: [{$ref: "#/components/parameters/ClusterId"}]
      responses:
        "200":
          description: Observed durable backup manifests, inferred WAL window, and isolated restore drill evidence
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BackupRecoverability"}
    post:
      operationId: createManualBackup
      parameters:
        - {$ref: "#/components/parameters/ClusterId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {type: {type: string, enum: [full, diff, incr], default: full}}
      responses:
        "202": {description: Manual backup job accepted}
        "409": {description: Backup policy or provider is unavailable}
...
components:
  securitySchemes:
    serviceAssertion:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable caller-generated key reused for every retry of the same mutation.
      schema: {type: string, minLength: 1, maxLength: 255}
    ClusterId:
      name: clusterId
      in: path
      required: true
      schema: {type: string, minLength: 1}
...
  schemas:
...
    BackupRecoverability:
      type: object
      required: [backups, recoveryWindow, confidence, isStale, blockers, drill]
      properties:
        backups:
          type: array
          items: {$ref: "#/components/schemas/ObservedBackup"}
        recoveryWindow:
          type: object
          required: [earliest, latest]
          properties:
            earliest: {type: string, format: date-time, nullable: true}
            latest: {type: string, format: date-time, nullable: true}
        confidence: {type: string, enum: [unknown, inferred, drill-verified]}
        isStale: {type: boolean}
        blockers: {type: array, items: {type: string}}
        drill:
          allOf:
            - {$ref: "#/components/schemas/RestoreDrill"}
          nullable: true
    ObservedBackup:
      type: object
      required: [id, type, status, startedAt, completedAt, recoverableUntil]
      properties:
        id: {type: string}
        type: {type: string, enum: [full, diff, incr]}
        status: {type: string, enum: [running, completed, failed]}
        startedAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time, nullable: true}
        recoverableUntil: {type: string, format: date-time, nullable: true}
    RestoreDrill:
      type: object
      required: [id, targetTime, completedAt, passed, evidenceDigest]
      properties:
        id: {type: string}
        targetTime: {type: string, format: date-time}
        completedAt: {type: string, format: date-time}
        passed: {type: boolean}
        evidenceDigest: {type: string, nullable: true}